+++ LibraryWebApi/Library.Application/UseCases/BookUseCases/DeleteBookUseCase.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using Library.Application.Exceptions;
 using Library.Domain.Interfaces;
 using Microsoft.AspNetCore.Mvc;
 
@@ -15,6 +16,13 @@ namespace Library.Application.UseCases.BookUseCases
 
         public async Task DeleteBook([FromRoute] int id)
         {
+            var existingBook = await _unitOfWork.Book.GetByIdAsync(id);
+
+            if (existingBook is null)
+            {
+                throw new EntityNotFoundException($"{existingBook} is not found in database.");
+            }
+
             await _unitOfWork.Book.DeleteAsync(id);
 
             await _unitOfWork.SaveChangesAsync();
